+++ public/src/components/channelManagement/epicTests/epicTestPreview.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Button, Dialog, makeStyles, Theme, Typography } from '@material-ui/core';
 import useOpenable from '../../../hooks/useOpenable';
 import EpicVariantPreview from './epicVariantPreview';
-import {EpicTest} from '../../../models/epic';
+import { EpicTest } from '../../../models/epic';
 
 const useStyles = makeStyles(({}: Theme) => ({
   dialog: {
