+++ public/src/components/channelManagement/epicTests/epicTestsForm.tsx
@@ -1,26 +1,31 @@
-import {FrontendSettingsType} from '../../../utils/requests';
-import React from 'react';
-import {TestsForm} from '../testsForm';
-import {getDefaultTest, getDefaultVariant} from './utils/defaults';
+import { FrontendSettingsType } from '../../../utils/requests';
+import { TestsForm } from '../testsForm';
+import { getDefaultTest, getDefaultVariant } from './utils/defaults';
 import {
   AMP_EPIC_CONFIG,
   APPLE_NEWS_EPIC_CONFIG,
   ARTICLE_EPIC_CONFIG,
   ARTICLE_EPIC_HOLDBACK_CONFIG,
   EpicEditorConfig,
-  LIVEBLOG_EPIC_CONFIG
+  LIVEBLOG_EPIC_CONFIG,
 } from '../helpers/shared';
-import { getEpicTestEditor } from './epicTestEditorNew';
-import {EpicTest} from '../../../models/epic';
+import { getEpicTestEditor } from './epicTestEditor';
+import { EpicTest } from '../../../models/epic';
 
-const createDefaultEpicTest = (epicEditorConfig: EpicEditorConfig) => (newTestName: string, newTestNickname: string): EpicTest => ({
+const createDefaultEpicTest = (epicEditorConfig: EpicEditorConfig) => (
+  newTestName: string,
+  newTestNickname: string,
+): EpicTest => ({
   ...getDefaultTest(),
   name: newTestName,
   nickname: newTestNickname,
   variants: epicEditorConfig.allowMultipleVariants ? [] : [{ ...getDefaultVariant() }],
 });
 
-const buildEpicTestsForm = (epicEditorConfig: EpicEditorConfig, settingsType: FrontendSettingsType) =>
+const buildEpicTestsForm = (
+  epicEditorConfig: EpicEditorConfig,
+  settingsType: FrontendSettingsType,
+) =>
   TestsForm(
     getEpicTestEditor(epicEditorConfig),
     settingsType,
