+++ public/src/components/channelManagement/helpers/shared.ts
@@ -24,7 +24,7 @@ export interface Test {
   articlesViewedSettings?: ArticlesViewedSettings;
   variants: Variant[];
   locations: Region[];
-  isNew?: boolean;  // true if test has not yet been POSTed to backend
+  isNew?: boolean; // true if test has not yet been POSTed to backend
 }
 
 export interface EpicEditorConfig {
