+++ public/src/components/channelManagement/stickyTopBar/testLockDetails.tsx
@@ -1,21 +1,11 @@
 import React from 'react';
 import { makeStyles, Typography } from '@material-ui/core';
-import { LockStatus } from './helpers/shared';
 
 const useStyles = makeStyles(() => ({
-  text: {
-    fontSize: '14px',
-    fontWeight: 500,
-    textTransform: 'uppercase',
-    letterSpacing: '1px',
+  lockDetailsText: {
+    alignSelf: 'flex-end',
   },
 }));
-
-interface StickyBottomBarDetailProps {
-  isInEditMode: boolean;
-  editedTestName: string | null;
-  lockStatus: LockStatus;
-}
 
 const formattedName = (email: string): string => {
   const nameArr: RegExpMatchArray | null = email.match(/^([a-z]*)\.([a-z]*).*@.*/);
@@ -52,29 +42,23 @@ const formattedTimestamp = (timestamp: string): string => {
   const paddedDay = String(day).padStart(2, '0');
   const monthName = MONTH_NAMES[date.getMonth()];
 
-  return `${monthName} ${paddedDay} - ${paddedhours}:${paddedMinutes}`;
+  return `${monthName} ${paddedDay} at ${paddedhours}:${paddedMinutes}`;
 };
 
-const StickyBottomBarDetail: React.FC<StickyBottomBarDetailProps> = ({
-  isInEditMode,
-  editedTestName,
-  lockStatus,
-}: StickyBottomBarDetailProps) => {
+interface TestLockDetailsProps {
+  email?: string;
+  timestamp?: string;
+}
+
+export const TestLockDetails: React.FC<TestLockDetailsProps> = ({
+  email,
+  timestamp,
+}: TestLockDetailsProps) => {
   const classes = useStyles();
-  let text = editedTestName == null ? '' : `- ${editedTestName}`;
+  if (email && timestamp) {
+    const text = `Locked by ${formattedName(email)}, since ${formattedTimestamp(timestamp)}`;
 
-  if (!isInEditMode && lockStatus.locked && lockStatus.email && lockStatus.timestamp) {
-    const name = formattedName(lockStatus.email);
-    const timestamp = formattedTimestamp(lockStatus.timestamp);
-
-    text = `- By ${name}, last live ${timestamp}`;
+    return <Typography className={classes.lockDetailsText}>{text}</Typography>;
   }
-
-  return (
-    <div>
-      <Typography className={classes.text}>{text}</Typography>
-    </div>
-  );
+  return null;
 };
-
-export default StickyBottomBarDetail;
