+++ public/src/components/channelManagement/testList.tsx
@@ -58,7 +58,6 @@ const TestList = <T extends Test>({
                           test={test}
                           isSelected={test.name === selectedTestName}
                           isEdited={!!test.lockStatus?.locked}
-                          onClick={(): void => {}}  // Cannot select a test while reordering
                         />
                       </div>
                     )}
